polishing
Former-commit-id: bace5ef1bf883efd739bf1a665c90e0004902b2a

+++ fortscale/presidio-output/presidio-output-processor/src/main/java/presidio/output/processor/spring/AlertServiceElasticConfig.java
@@ -28,7 +28,7 @@ public class AlertServiceElasticConfig {
     private int eventsLimit;
 
     @Value("${indicators.contribution.limit.to.classification}")
-    private int contributionLimit;
+    private double contributionLimit;
 
     @Autowired
     private AlertClassificationService alertClassificationService;
